requirements.txt
asgiref==3.8.1
Django==4.2.18
djangorestframework==3.15.2
pillow==11.1.0
python-decouple==3.8
qrcode==8.0
sqlparse==0.5.3
typing-extensions==4.12.2
